[add] アルバムに公開期限を付加
