+++ app/assets/stylesheets/foundation/_variables_copy.sass
@@ -1,0 +1,78 @@
+// ==============================================
+// Color Map
+// ==============================================
+
+// site colors
+// ------------------------------------
+$main-color:  var(--main-color)
+$accent-color: var(--accent-color)
+
+// font weights
+// ------------------------------------
+$fontWeights: ('normal': 400,'bold': 700)
+// fontSize
+// @example
+//  font-size: map-get($fontSizes, 'm')
+$fontSizes: ('xs':  1.0rem, 's': 1.2rem,'m': 1.5rem, 'l': 1.8rem, 'xl': 2.0rem, 'xxl': 2.4rem)
+// layoutColors
+// @example
+//  background-color: map-get($layoutColors, 'main')
+// ------------------------------------
+$layoutColors: ( 'base': #f7f7f7, 'main': $main-color, 'accent': $accent-color, 'black': #000, 'richBlack': #333, 'darkGray': #666, 'gray': #bbb, 'lightGray': #efefef, 'white': #fff, 'blue': #04a9a4, 'orange': #ff8c00, 'red': #f33)
+// fontColors
+// ------------------------------------
+$fontColors: ( 'main': $main-color, 'accent': $accent-color, 'black': #333, 'darkGray': #666, 'gray': #bbb, 'white': #fff, 'red': #ea3033, 'blue': #3664c9, 'orange': #ff8c00, 'gold': #DCB769)
+
+// svgColors
+// ------------------------------------
+$svgColors: ( 'main': $main-color, 'black': #333, 'darkGray': #666, 'gray': #bbb, 'white': #fff, 'gold': #DCB769)
+// svgSize
+// @example
+//  svg-size: map-get($svgSizes, 'm')
+$svgSizes: ( 'xs': 18px, 's': 21px, 'm': 24px, 'l': 34px, 'xl': 75px)
+
+// button colors
+// ------------------------------------
+$buttonColors: ( 'main': $main-color, 'primary': #3664c9, 'danger': #93110a, 'negative': #999)
+
+// SNS Symbol Colors
+// ------------------------------------
+$snsColors: ( 'facebook': #305097, 'twitter': #55acee, 'google': #dd4b39, 'line': #00C300, 'github': #333333, 'hatena': #00A2E4, 'pocket': #ED4156)
+
+// Ranking Label Colors
+// ------------------------------------
+$rankingColors: ( '1st': #d9b340, '2nd': #6B7380, '3rd': #a15326)
+
+// ==============================================
+// settings
+// ==============================================
+
+// コンテンツの幅
+// ------------------------------------
+$container-s: 780px
+$container-m: 1100px
+$container-l: 1440px
+
+// 右カラムの最大幅
+// ------------------------------------
+$contents-rightSideWidth: 300px
+
+// エレメント間のマージン(SP)
+$element-margin-sp: 20px
+
+// 左右のpadding
+$element-padding-sp: 3.8%
+
+// エレメント間のマージン(PC)
+$element-margin-pc: 30px
+
+// z-index値
+// @example
+//  z-index: $layerLevel-header + 100
+// ------------------------------------
+// 10000 〜
+$layerLevel-overlay: 10000
+// 8000 〜 9999
+$layerLevel-drawer: 8000
+// 5000 〜 7999
+$layerLevel-header: 5000
