+++ app/controllers/front/albums_controller.rb
@@ -1,7 +1,7 @@
 class Front::AlbumsController < Front::ApplicationController
   before_action :login_required
   before_action :validate_user, only: [:show]
-  before_action :close_album, only: [:index]
+  before_action :close_album, only: [:index, :show]
   before_action :overdue, only: [:show]
 
   def index
@@ -15,6 +15,7 @@ class Front::AlbumsController < Front::ApplicationController
   end
 
   private
+
     # ログイン中のユーザーを検証
     def validate_user
       album = Album.find(params[:id])
@@ -23,23 +24,5 @@ class Front::AlbumsController < Front::ApplicationController
         redirect_to front_root_url, flash: { danger: 'アクセス権限がありません' }
       end
     end
-    # 公開期間終了したアルバムを非公開にする
-    def close_album
-      albums = current_user.albums.where.not(status: 'close')
-      albums.each do |album|
-        deadline = album.created_at.to_date + album.open_period
-        if deadline < Date.today
-          album.update!(status: 'closed')
-        end
-      end
-    end
-    # 公開期間の徒過を検証
-    def overdue
-      album = Album.find(params[:id])
-      deadline = album.created_at.to_date + album.open_period
-      if deadline < Date.today
-        redirect_to front_albums_url, flash: { danger: 'アルバムの公開期間は終了しています' } 
-      end
-    end
-    
+
 end
