+++ app/controllers/front/histories_controller.rb
@@ -1,11 +1,14 @@
 class Front::HistoriesController < Front::PhotosController
+  before_action :close_album, only: [:index, :show]
+  before_action :overdue, only: [:show]
+  
   def index
-    @photos = Photo.re_selected.joins(album: :user).merge(User.where(id: current_user.id)).reverse_order
+    @photos = Photo.re_selected.joins(album: :user).merge(User.where(id: current_user.id)).merge(Album.where.not(status: 'closed')).reverse_order
   end
 
   def show
-    @photos = Photo.complete.joins(album: :user).merge(User.where(id: current_user.id)).merge(Album.where(id: params[:id])).reverse_order
-    .or(Photo.re_selected.joins(album: :user).merge(User.where(id: current_user.id)).merge(Album.where(id: params[:id])).reverse_order)
+    @photos = Photo.complete.joins(album: :user).merge(User.where(id: current_user.id)).merge(Album.where(id: params[:id])).merge(Album.where.not(status: 'closed')).reverse_order
+    .or(Photo.re_selected.joins(album: :user).merge(User.where(id: current_user.id)).merge(Album.where(id: params[:id])).merge(Album.where.not(status: 'closed')).reverse_order)
   end
 
   def update
