--- app/javascript/components/data_modal/info_icon/index.jsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import Style from './style.sass'
-
-export default class InfoIcon extends React.Component {
-
-  // 表示処理
-  render () {
-
-    return(
-      <div className={Style.InfoIcon}>
-      </div>
-    );
-  }
-}
