--- app/javascript/components/data_modal/info_icon/style.sass
@@ -1,39 +1,0 @@
-// ==============================================
-// Info Icon
-// ==============================================
-@import '../../../../assets/stylesheets/foundation/variables'
-
-.InfoIcon
-
-  width: 80px
-  height: 80px
-  border-width: 4px
-  border-style: solid
-  border-radius: 50%
-  padding: 0
-  position: relative
-  box-sizing: content-box
-  margin: 20px auto
-  border-color: map-get($layoutColors, 'blue')
-
-  &::before
-    content: ''
-    position: absolute
-    width: 5px
-    height: 29px
-    left: 50%
-    bottom: 17px
-    border-radius: 2px
-    margin-left: -2px
-    background-color: map-get($layoutColors, 'blue')
-
-  &::after
-    content: ''
-    position: absolute
-    width: 7px
-    height: 7px
-    border-radius: 50%
-    margin-left: -3px
-    top: 19px
-    background-color: map-get($layoutColors, 'blue')
-    left: 50%
