+++ app/javascript/components/loading/style.sass
@@ -1,14 +1,14 @@
-@import '../../../assets/stylesheets/foundation/variables'
+// @import '../../../assets/stylesheets/foundation/variables'
 .Loading
   position: fixed
   top: 0
   left: 0
   right: 0
   bottom: 0
-  background-color: rgba(map-get($layoutColors, 'black'), .3)
+  background-color: rgba(0, 0, 0, .3)
   display: flex
   justify-content: center
   align-items: center
   font-size: 2.4rem
-  color: map-get($fontColors, 'white')
-  z-index: $layerLevel-overlay
+  color: #ffffff
+  z-index: 10000
