+++ app/models/album.rb
@@ -51,4 +51,5 @@ class Album < ApplicationRecord
   #----------------------------------------
   #  ** Methods **
   #----------------------------------------
+  
 end
