+++ app/views/admin/albums/index.html.slim
@@ -9,6 +9,7 @@
           th.l-table__header クライアント名
           th.l-table__header 登録枚数
           th.l-table__header 公開設定
+          th.l-table__header 作成日
           th.l-table__header 公開期限
           th.l-table__header
           th.l-table__header
@@ -20,6 +21,7 @@
             td.l-table__data = album.user.name
             td.l-table__data = album.photos.count
             td.l-table__data = album.status_i18n
+            td.l-table__data = album.created_at.to_date
             td.l-table__data = album.created_at.to_date + album.open_period
             td.l-table__data = react_component 'album_form', user: album.user.as_json, categories: Album.categories.keys.map {|k| [I18n.t("enums.album.category.#{k}"),k]}.as_json, album: album.as_json
             td.l-table__data = link_to '削除', admin_album_path(album.id), method: :delete, data: { confirm: '削除してよろしいですか？' }, class: 'p-albums__delete'
