+++ app/views/front/layouts/_partials/_header.html.slim
@@ -1,9 +1,7 @@
-header.l-dashboard__header
-
-  .l-dashboard__headerTabs
-    = link_to 'メインメニュー', front_root_path
-
-  .l-dashboard__headerMenus
-    = link_to 'プロフィール編集', edit_front_user_path(current_user), class: 'l-dashboard__headerMenu'
-    = link_to 'ダウンロードリスト', front_photos_path, class: 'l-dashboard__headerMenu'
-    = link_to 'ログアウト', logout_path, class: 'l-dashboard__headerMenu'
+header.l-header
+  .header__inner
+    .l-header__nav
+      = link_to 'メインメニュー', front_root_path
+      = link_to 'プロフィール編集', edit_front_user_path(current_user), class: 'l-dashboard__headerMenu'
+      = link_to 'ダウンロードリスト', front_photos_path, class: 'l-dashboard__headerMenu'
+      = link_to 'ログアウト', logout_path, class: 'l-dashboard__headerMenu'
